Add admin-RPC SO_PEERCRED gate and widen RPC peer UID to int64_t
Open-source portion of rstudio-pro#10941. Replaces uid_t->int narrowing
on RPC peer-uid fields with int64_t so the nobody sentinel (4294967295)
is no longer misread as -1, and adds an SO_PEERCRED-driven admin-auth
gate for session RPC endpoints registered with allowUserAccess=false.

- ServerSessionRpc.cpp: thread requireAdminAuth through validationHandler
  / addHandler / addHttpProxyHandler. When set, the handler rejects any
  caller whose peer uid is not root or the configured server-user with
  a 403. Server-user uid is resolved at session_rpc::initialize() via
  User::getUserFromIdentifier and cached as an int64_t.

- ServerSessionRpcOverlay.cpp: widen the overlay typedef to match.

- SecureKeyFile.{cpp,hpp}: always try system paths first (the service
  account owns those and can read them), only fall back to the user
  cache when no system key exists and the caller is unprivileged. Add
  systemKeyFilePath() to expose the system path without reading or
  creating the file.

- ServerMain.cpp: in --verify-installation mode, skip secure-cookie /
  session-rpc init when the key file exists but is unreadable by the
  caller. Uses systemKeyFilePath() to compute the path without reading.

- Request.hpp, SessionOptions.{cpp,hpp},
  SessionLocalStreamHttpConnectionListener.hpp: widen remoteUid /
  limitRpcClientUid plumbing to int64_t to match.

+++ src/cpp/session/SessionOptions.cpp
@@ -425,7 +425,7 @@ core::ProgramStatus Options::read(int argc, char * const argv[], std::ostream& o
    std::string limitUid = core::system::getenv(kRStudioLimitRpcClientUid);
    if (!limitUid.empty())
    {
-      limitRpcClientUid_ = core::safe_convert::stringTo<int>(limitUid, -1);
+      limitRpcClientUid_ = core::safe_convert::stringTo<int64_t>(limitUid, -1);
       core::system::unsetenv(kRStudioLimitRpcClientUid);
    }
